+++ download.sh
@@ -10,11 +10,12 @@ export VENV_DIR=$PWD/tmp
 PIP=pip
 
 # cleanup
-/bin/rm installer/cache/*
+/bin/rm installer/cache/* installer/cache/.python2 >/dev/null 2>&1
 
 if [ -e $VENV_DIR/bin/python2 ]; then
     # It seems that the latest pip does not work for python 2...
     echo "Do not upgrade pip / setuptools for python 2"
+    touch installer/cache/.python2
 else
     # Update pip / setuptools
     $PIP install -U pip