+++ installer/install.sh
@@ -3,7 +3,11 @@
 export VENV_DIR=${VENV_DIR:-/opt/python-env}
 
 # create venv
-./create-venv.sh "$VENV_DIR" $@ || exit 1
+if [ -e ./cache/.python2 ]; then
+    ./create-venv.sh "$VENV_DIR" -2 || exit 1
+else
+    ./create-venv.sh "$VENV_DIR" || exit 1
+fi
 
 # activate venv
 . $VENV_DIR/bin/activate || exit 1
@@ -13,12 +17,14 @@ echo "pip = $(command -v pip)"
 
 INSTALL="pip install -U --no-index --disable-pip-version-check --find-links=./cache" # -v
 
-# install pip
-#python $PIP_WHL/pip install -U pip*.whl
-$INSTALL pip
+if [ ! -e ./cache/.python2 ]; then  # Do not update pip / setuptools for python 2
+    # update pip
+    #python $PIP_WHL/pip install -U pip*.whl
+    $INSTALL pip
 
-# install setuptools
-$INSTALL setuptools
+    # update setuptools
+    $INSTALL setuptools
+fi
 
 # install packages
 $INSTALL -r ./requirements.txt -v