+++ ChangeLog.md
@@ -1,9 +1,10 @@
 # ChangeLog
 
-## 2020/xx/xx : v1.0.1
+## 2020/08/22 : v1.0.1
 
-- Default to use python 3 
+- Default to use python 3
+- Do not upgrade pip / setuptools for python 2 
 
-## 2020/7/12 : v1.0.0
+## 2020/07/12 : v1.0.0
 
 - Initial release